third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/line-height-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'line-height' to a length: -3.14em
  assert_class_string: relative lengths must compute to a CSSUnitValue expected "[object CSSUnitValue]" but got "[object CSSKeywordValue]"
[FAIL] Can set 'line-height' to a number: -3.14
  assert_class_string: relative lengths must compute to a CSSUnitValue expected "[object CSSUnitValue]" but got "[object CSSKeywordValue]"
[FAIL] Can set 'line-height' to a percent: -3.14%
  assert_class_string: relative lengths must compute to a CSSUnitValue expected "[object CSSUnitValue]" but got "[object CSSKeywordValue]"
Harness: the test ran to completion.
